+++ lib/module/appsflyer/affise_module_appsflyer_api.dart
@@ -1,0 +1,3 @@
+abstract class AffiseModuleAppsFlyerApi {
+  void logEvent(String eventName, Map<String, dynamic> eventValues);
+}