+++ lib/module/link/affise_module_link_api.dart
@@ -1,0 +1,5 @@
+import 'affise_link_callback.dart';
+
+abstract class AffiseModuleLinkApi {
+  void resolve(String url, AffiseLinkCallback callback);
+}