+++ lib/module/subscription/export.dart
@@ -5,3 +5,4 @@ export 'affise_products_result.dart';
 export 'affise_purchased_info.dart';
 export 'affise_product_subscription_detail.dart';
 export 'time_unit_type.dart';
+export 'affise_module_subscription_api.dart';